+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - PR #49 - Add CuPy Module for convolve2d and correlate2d
 - PR #51 - Add CuPy Module for lombscargle, along with tests/benchmarks
 - PR #62 - Add CuPy Module for 1d convolve and correlate, along with tests/benchmarks
+- PR #66 - Add CuPy Module for 2d upfirdn, along with tests/benchmarks
 
 ## Improvements
 - PR #40 - Ability to specify time/freq domain for resample.